% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_mapfiles}
\alias{get_mapfiles}
\title{Get map file paths}
\usage{
get_mapfiles(d = options()$ukbpppdir)
}
\arguments{
\item{d}{Directory (defaults to bc4 location)}
}
\value{
array of paths for each map file chromosome
}
\description{
Get map file paths
}
