% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_snpid_from_rsid}
\alias{get_snpid_from_rsid}
\title{Get variant information based on rsid lookup}
\usage{
get_snpid_from_rsid(map_files = get_mapfiles(), rsid)
}
\arguments{
\item{rsid}{vector of variants by rsid}

\item{map_dir}{path to the \code{Metadata/SNP RSID maps} directory}
}
\value{
data frame
}
\description{
Provide a path to the \code{Metadata/SNP RSID maps} directory and lookup the snp IDs by rsid
}
