% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{protein_info}
\alias{protein_info}
\title{Get protein information}
\usage{
protein_info(d = options()$ukbpppdir)
}
\arguments{
\item{d}{base directory (defaults to bc4 location)}
}
\value{
data frame
}
\description{
Adds the tarball file name column too
}
