% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_pgwas}
\alias{query_pgwas}
\title{Query UKB-PPP downloaded files}
\usage{
query_pgwas(tarfile, variants)
}
\arguments{
\item{tarfile}{Path to a tarball e.g. \code{/path/to/African/A1BG_P04217_OID30771_v1_Inflammation_II.tar}}

\item{variants}{Output from \code{get_snpid_from_chrpos} or \code{get_snpid_from_rsid}}
}
\value{
data frame
}
\description{
The files are arranged according to https://www.synapse.org/#!Synapse:syn51364943/files/.
The \code{metadata/} directory has a mapping of rsid and chr/pos in build hg19 and hg38.
The \code{UKB-PPP pGWAS summary statistics} directory has a directory for each ancestry
with complete summary statistics for ~3k proteins in each directory. Each GWAS result is
split into separate chromosome files and then tarballed.
}
\details{
This function takes a data frame of variants and extracts them from a specified tarball file
}
