% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{standardise}
\alias{standardise}
\title{Standardise data to have alphabetical effect allele ordering and variant id}
\usage{
standardise(
  d,
  ea = "ea",
  oa = "oa",
  eaf = "eaf",
  beta = "beta",
  chr = "chr",
  pos = "pos"
)
}
\arguments{
\item{d}{data frame}

\item{ea}{effect allele column name}

\item{oa}{other allele column name}

\item{eaf}{effect allele frequency column name}

\item{beta}{beta column name}

\item{chr}{chr column name}

\item{pos}{pos column name}
}
\value{
data frame
}
\description{
Generates variant id chr:pos_a1_a2 with a1 and a2 oriented alphabetically
flips eaf and beta where necessary
}
